Fixed issues raised by compiler warnings.

+++ CvGameCoreDLL/CvCity.cpp
@@ -3607,7 +3607,7 @@ UnitTypes CvCity::getConscriptUnit(bool bFree) const
 	std::vector<UnitTypes> aeConscriptUnits = getConscriptUnits();
 	std::vector<UnitTypes> aePossibleUnits;
 	
-	for (int iI = 0; iI < aeConscriptUnits.size(); ++iI)
+	for (int iI = 0; iI < (int)aeConscriptUnits.size(); ++iI)
 	{
 		if (bFree || canConscript(aeConscriptUnits[iI]))
 		{
@@ -3628,7 +3628,7 @@ std::wstring CvCity::getConscriptText() const
 {
 	std::vector<UnitTypes> aeConscriptUnits = getConscriptUnits();
 	std::wstring szConscriptText = std::wstring();
-	for (int iI = 0; iI < aeConscriptUnits.size(); ++iI)
+	for (int iI = 0; (int)iI < aeConscriptUnits.size(); ++iI)
 	{
 		if (iI > 0)
 		{
@@ -3729,7 +3729,7 @@ bool CvCity::canConscript() const
 
 	bool bFound = false;
 	std::vector<UnitTypes> aeConscriptUnits = getConscriptUnits();
-	for (int iI = 0; iI < aeConscriptUnits.size(); ++iI)
+	for (int iI = 0; iI < (int)aeConscriptUnits.size(); ++iI)
 	{
 		if (canConscript(aeConscriptUnits[iI]))
 		{
